# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Associació cultural de Montcada",
        "version": "0.1"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Display a listing of the users",
                "description": "Display a listing of the users.",
                "operationId": "indexUser",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Store a user in database",
                "description": "Store a user in database.",
                "operationId": "storeUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Josep"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "example@example.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/user/{userId}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Display a specific user",
                "description": "Display a specific user.",
                "operationId": "showUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id of user that needs to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/users/{userId}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user in database",
                "description": "Update a user in database.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id of user that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Josep"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "example@example.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "12345678"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "format": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user in database",
                "description": "Delete a user in database.",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "Id of user that needs to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}